server/requirements.txt
anyio==3.6.1
beanie==1.26.0
click==8.1.3
fastapi==0.85.0
gunicorn==20.1.0
h11==0.14.0
idna==3.4
lazy-model==0.2.0
motor==3.4.0
pydantic==1.10.2
pymongo==4.5.0
python-dateutil==2.8.2
pytz==2024.1
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
toml==0.10.2
tzdata==2024.1
uvicorn==0.18.3
email-validator
pydantic[email]
passlib
python-jose
python-multipart
shapely
geopandas